refactor : resetTransform 커스텀 훅 분리
